requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==45.0.6
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
duckdb==1.3.2
duckdb_engine==0.17.0
emoji==2.14.1
executing==2.2.0
filetype==1.2.0
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-ollama==0.3.7
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.17
lxml==6.0.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.3.2
olefile==0.47
ollama==0.5.3
openai==1.101.0
orjson==3.11.2
packaging==25.0
parso==0.8.5
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
RapidFuzz==3.13.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.11.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unstructured==0.18.14
unstructured-client==0.42.3
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.17.3
yarl==1.20.1
zstandard==0.24.0
fastapi==0.118.0
uvicorn[standard]==0.34.0
websockets==14.1
chromadb==0.4.15
sentence-transformers==3.3.1
pydantic-settings==2.10.1
